If test requirements are not saticfied, do not compile test code. Before this patch test code was not executed, but compiled in any case.
The thing is that test code can use definitions, which existence is tested in 'Require' section.

E.g., on Debian 6 FITRIM macro is not defined, so test ioctl.IoctlFITRIM is unsupported.
But code of the test uses FITRIM macro, so it failed to be compiled if definition of that macro is absent.