Add support for <RunIf> tag for test and testsets. Make tests use it when it is possible.
<RunIf> tag contains (runtime-evaluated) condition for run test or (if specified for testset) all tests in the testset.
Currently, that condition is evaluated at the very beginning of the test. If it is failed, test is marked as unsupported;
'message' attribute may be given for specify reason of that.
But that condition may be evaluated for decide, whether test should be included into testset. Such evaluation will help to shortening time needed
for tests execution, because it is not needed to mkfs/mount filesystem before test, which will not be run.

Make some ext4-specific tests use "mount-oriented='false'", "per-mount-options='false'" attributes.